// frontend/src/api/model/outQuesitonOutType.ts
/**
 * Generated by orval v6.17.0 🍺
 * Do not edit manually.
 * TT API
 * Team tuner API documentation
 * OpenAPI spec version: 1.0
 */

export type OutQuesitonOutType =
  (typeof OutQuesitonOutType)[keyof typeof OutQuesitonOutType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OutQuesitonOutType = {
  extraversion: 'extraversion',
  agreeableness: 'agreeableness',
  conscientious: 'conscientious',
  neuroticism: 'neuroticism',
  openness: 'openness',
} as const;